+++ lib/options.dart
@@ -1,7 +1,0 @@
-// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-@Deprecated(
-    'Will be removed in a later version of DartDoc. Use /src/dartdoc_options.dart.')
-export 'package:dartdoc/src/dartdoc_options.dart';