+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Remove the deprecated `templates-dir` option.
 * Remove the deprecated `nodoc` option.
 * Remove the deprecated `include-external` option.
+* Refactor 404 error page to use div instead of p for search form.
 
 ## 8.3.4
 * The URL for category pages now uses _category name_ instead of