+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Remove the deprecated `nodoc` option.
 * Remove the deprecated `include-external` option.
 * Refactor 404 error page to use div instead of p for search form.
+* Require analyzer 8.1.1 APIs.
+* Many internal migrations for analyzer APIs.
 
 ## 8.3.4
 * The URL for category pages now uses _category name_ instead of